doc/authoring.md: avoid unreliable parsing of `ls`
Parsing `ls` is notoriously fragile and should be avoided when possible.
In this case, `ls` was an alias to `ls -h` for me, resulting in `3.0K`
as a file size instead of a number of bytes.

Replace it with the reliable `stat` instead, which is part of coreutils
just like `printf` used in the rest of these instructions.

Signed-off-by: Eric Engestrom <eric.engestrom@imgtec.com>